+++ payload/game/ui/VotingBackPage.hh
@@ -59,8 +59,8 @@ private:
         Handler(VotingBackPage &page);
         ~Handler();
 
-        void onReceivePulse(u32 playerId) override;
-        void onReceiveInfo(u32 playerId, s32 course, u32 selectedPlayer, u32 character,
+        void onReceivePulse(s8 playerId) override;
+        void onReceiveInfo(s8 playerId, s32 course, u32 selectedPlayer, u32 character,
                 u32 vehicle) override;
 
     private:
